pyproject.toml
[tool.poetry]
name = "iniciativa_dados_abertos"
version = "0.1.0"
description = ""
authors = ["Rodrigo Torres <ratorres7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^2.11.2"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
